Rename BlockIndexPanels component as BlockPanelLinks
